editor/libeditor/tests/test_nsITableEditor_deleteTableCellContents.html

<!DOCTYPE>
<html>
<head>
  <title>Test for nsITableEditor.deleteTableCellContents()</title>
  <script src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" href="/tests/SimpleTest/test.css">
</head>
<body>
<div id="display">
</div>
<div id="content" contenteditable>out of table<table><tr><td>default content</td></tr></table></div>
<pre id="test">
</pre>

<script class="testbody" type="application/javascript">

SimpleTest.waitForExplicitFinish();
SimpleTest.waitForFocus(function() {
  let editor = document.getElementById("content");
  let selection = document.getSelection();

  function checkInputEvent(aEvent, aDescription) {
    ok(aEvent instanceof InputEvent,
       `"input" event should be dispatched with InputEvent interface ${aDescription}`);
    is(aEvent.cancelable, false,
       `"input" event should be never cancelable ${aDescription}`);
    is(aEvent.bubbles, true,
       `"input" event should always bubble ${aDescription}`);
    is(aEvent.inputType, "deleteContent",
       `inputType should be "deleteContent" ${aDescription}`);
     is(aEvent.data, null,
       `data should be null ${aDescription}`);
    is(aEvent.dataTransfer, null,
       `data should be null ${aDescription}`);
 }

  let inputEvents = [];
  function onInput(aEvent) {
    inputEvents.push(aEvent);
  }
  editor.addEventListener("input", onInput);

  selection.collapse(editor.firstChild, 0);
  inputEvents = [];
  getTableEditor().deleteTableCellContents();
  is(editor.innerHTML, "out of table<table><tbody><tr><td>default content</td></tr></tbody></table>",
     "nsITableEditor.deleteTableCellContents() should do nothing if selection is not in <table>");
  is(inputEvents.length, 0,
     'No "input" event should be fired when a call of nsITableEditor.deleteTableCellContents() does nothing');

  selection.removeAllRanges();
  try {
    inputEvents = [];
    getTableEditor().deleteTableCellContents();
    ok(false, "getTableEditor().deleteTableCellContents() without selection ranges should throw exception");
  } catch (e) {
    ok(true, "getTableEditor().deleteTableCellContents() without selection ranges should throw exception");
    is(inputEvents.length, 0,
       'No "input" event should be fired when nsITableEditor.deleteTableCellContents() causes exception due to no selection range');
  }

  selection.removeAllRanges();
  editor.innerHTML = "<table>" +
                       '<tr><td id="select">cell1-1</td><td>cell1-2</td></tr>' +
                       "<tr><td>cell2-1</td><td>cell2-2</td></tr>" +
                     "</table>";
  editor.focus();
  inputEvents = [];
  let range = document.createRange();
  range.selectNode(document.getElementById("select"));
  selection.addRange(range);
  getTableEditor().deleteTableCellContents();
  is(editor.innerHTML, "<table><tbody>" +
                         '<tr><td id="select"><br></td><td>cell1-2</td></tr>' +
                         "<tr><td>cell2-1</td><td>cell2-2</td></tr>" +
                       "</tbody></table>",
     "nsITableEditor.deleteTableCellContents() should replace the selected cell's text with <br> element");
  is(inputEvents.length, 1,
     'Only one "input" event should be fired when all text in a cell is selected');
  checkInputEvent(inputEvents[0], "when all text in a cell is selected");

  selection.removeAllRanges();
  editor.innerHTML = "<table>" +
                       '<tr><td id="select"><ul><li>list1</li></ul></td><td>cell1-2</td></tr>' +
                       "<tr><td>cell2-1</td><td>cell2-2</td></tr>" +
                     "</table>";
  editor.focus();
  inputEvents = [];
  range = document.createRange();
  range.selectNode(document.getElementById("select"));
  selection.addRange(range);
  getTableEditor().deleteTableCellContents();
  is(editor.innerHTML, "<table><tbody>" +
                         '<tr><td id="select"><br></td><td>cell1-2</td></tr>' +
                         "<tr><td>cell2-1</td><td>cell2-2</td></tr>" +
                       "</tbody></table>",
     "nsITableEditor.deleteTableCellContents() should replace the selected cell's <ul> element with <br> element");
  is(inputEvents.length, 1,
     'Only one "input" event should be fired when <ul> element in a cell is selected');
  checkInputEvent(inputEvents[0], "when <ul> element in a cell is selected");

  selection.removeAllRanges();
  editor.innerHTML = "<table>" +
                       '<tr><td id="select1">cell1-1</td><td>cell1-2</td></tr>' +
                       '<tr><td id="select2">cell2-1</td><td>cell2-2</td></tr>' +
                     "</table>";
  editor.focus();
  inputEvents = [];
  range = document.createRange();
  range.selectNode(document.getElementById("select1"));
  selection.addRange(range);
  range = document.createRange();
  range.selectNode(document.getElementById("select2"));
  selection.addRange(range);
  getTableEditor().deleteTableCellContents();
  is(editor.innerHTML, "<table><tbody>" +
                         '<tr><td id="select1"><br></td><td>cell1-2</td></tr>' +
                         '<tr><td id="select2"><br></td><td>cell2-2</td></tr>' +
                       "</tbody></table>",
     "nsITableEditor.deleteTableCellContents() should replace the selected 2 cells' text with <br> element");
  is(inputEvents.length, 1,
     'Only one "input" event should be fired when 2 cell elements are selected');
  checkInputEvent(inputEvents[0], "when 2 cell elements are selected");

  selection.removeAllRanges();
  editor.innerHTML = "<table>" +
                       '<tr><td id="select1">cell1-1</td><td id="select2">cell1-2</td></tr>' +
                       '<tr><td id="select3">cell2-1</td><td id="select4">cell2-2</td></tr>' +
                     "</table>";
  editor.focus();
  inputEvents = [];
  range = document.createRange();
  range.selectNode(document.getElementById("select1"));
  selection.addRange(range);
  range = document.createRange();
  range.selectNode(document.getElementById("select2"));
  selection.addRange(range);
  range = document.createRange();
  range.selectNode(document.getElementById("select3"));
  selection.addRange(range);
  range = document.createRange();
  range.selectNode(document.getElementById("select4"));
  selection.addRange(range);
  getTableEditor().deleteTableCellContents();
  is(editor.innerHTML, "<table><tbody>" +
                         '<tr><td id="select1"><br></td><td id="select2"><br></td></tr>' +
                         '<tr><td id="select3"><br></td><td id="select4"><br></td></tr>' +
                       "</tbody></table>",
     "nsITableEditor.deleteTableCellContents() should replace the selected 4 cells' text with <br> element");
  is(inputEvents.length, 1,
     'Only one "input" event should be fired when 4 cell elements are selected');
  checkInputEvent(inputEvents[0], "when 4 cell elements are selected");

  selection.removeAllRanges();
  editor.innerHTML = "<table>" +
                       '<tr><td id="select" rowspan="2">cell1-1</td><td>cell1-2</td></tr>' +
                       "<tr><td>cell2-2</td></tr>" +
                     "</table>";
  editor.focus();
  inputEvents = [];
  range = document.createRange();
  range.selectNode(document.getElementById("select"));
  selection.addRange(range);
  getTableEditor().deleteTableCellContents();
  is(editor.innerHTML, "<table><tbody>" +
                         '<tr><td id="select" rowspan="2"><br></td><td>cell1-2</td></tr>' +
                         "<tr><td>cell2-2</td></tr>" +
                       "</tbody></table>",
     "nsITableEditor.deleteTableCellContents() should replace the selected cell's text with <br> element (even if the cell is row-spanning)");
  is(inputEvents.length, 1,
     'Only one "input" event should be fired when a cell element are selected (even if the cell is row-spanning)');
  checkInputEvent(inputEvents[0], "when a cell element are selected (even if the cell is row-spanning)");

  selection.removeAllRanges();
  editor.innerHTML = "<table>" +
                       '<tr><td id="select" colspan="2">cell1-1</td></tr>' +
                       "<tr><td>cell2-1</td><td>cell2-2</td></tr>" +
                     "</table>";
  editor.focus();
  inputEvents = [];
  range = document.createRange();
  range.selectNode(document.getElementById("select"));
  selection.addRange(range);
  getTableEditor().deleteTableCellContents();
  is(editor.innerHTML, "<table><tbody>" +
                          '<tr><td id="select" colspan="2"><br></td></tr>' +
                          "<tr><td>cell2-1</td><td>cell2-2</td></tr>" +
                        "</tbody></table>",
     "nsITableEditor.deleteTableCellContents() should replace the selected cell's text with <br> element (even if the cell is column-spanning)");
  is(inputEvents.length, 1,
     'Only one "input" event should be fired when a cell element are selected (even if the cell is column-spanning)');
  checkInputEvent(inputEvents[0], "when a cell element are selected (even if the cell is column-spanning)");

  selection.removeAllRanges();
  editor.innerHTML = "<table>" +
                       '<tr><td id="select">cell1-1</td><td>cell1-2</td></tr>' +
                       "<tr><td>cell2-1</td><td>cell2-2</td></tr>" +
                     "</table>";
  editor.focus();
  editor.scrollTop; // Requires layout information.
  inputEvents = [];
  selection.setBaseAndExtent(document.getElementById("select").firstChild, 0,
                             document.getElementById("select").firstChild, 0);
  getTableEditor().deleteTableCellContents();
  is(editor.innerHTML, "<table><tbody>" +
                         '<tr><td id="select"><br></td><td>cell1-2</td></tr>' +
                         "<tr><td>cell2-1</td><td>cell2-2</td></tr>" +
                       "</tbody></table>",
     "nsITableEditor.deleteTableCellContents() should replace a cell's text with <br> element when the cell contains selection range");
  is(inputEvents.length, 1,
     'Only one "input" event should be fired when the cell contains selection range');
  checkInputEvent(inputEvents[0], "when the cell contains selection range");

  editor.removeEventListener("input", onInput);

  SimpleTest.finish();
});

function getTableEditor() {
  var editingSession = SpecialPowers.wrap(window).docShell.editingSession;
  return editingSession.getEditorForWindow(window).QueryInterface(SpecialPowers.Ci.nsITableEditor);
}

</script>
</body>

</html>
